refactor(tui): extract pure model picker helpers to dialog-model-utils.ts
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)
